+++ src/cinder/Json.cpp
@@ -26,7 +26,7 @@
 
 
 #include <boost/algorithm/string.hpp>
-#include <json/json.h>
+#include "jsoncpp/json.h"
 
 #include "cinder/Json.h"
 #include "cinder/Stream.h"
